refactor(scully): efactor contextRenderPlugin into reusable pieces to prepare other uses of it
