+++ libs/scully/src/lib/renderPlugins/content-render-utils/getIdAttrName.ts
@@ -1,0 +1,13 @@
+export function getIdAttrName(attrs: string): string {
+  try {
+    return (
+      attrs &&
+      attrs
+        .split(' ')
+        .find((at: string) => at.trim().startsWith('_ngcontent'))
+        .split('=')[0]
+    );
+  } catch {
+    return '6';
+  }
+}
