+++ libs/scully/tsconfig.json
@@ -18,6 +18,6 @@
     "typeRoots": [],
     "allowSyntheticDefaultImports": true
   },
-  "files": ["src/index.ts", "src/scully.ts", "./src/lib/utils/handlers/processRoutes.ts"],
+  "files": ["src/index.ts", "src/scully.ts"],
   "exclude": ["../../dist/**/*"]
 }
